+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:easy_permission_validator/easy_permisson_validator.dart';
+import 'package:easy_permission_validator/easy_permission_validator.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 void main() => runApp(MyApp());
 
@@ -37,7 +38,7 @@ class _MyHomePageState extends State<MyHomePage> {
             IconButton(
               icon: Icon(Icons.camera),
               iconSize: 90.0,
-              onPressed: () => _camera(),
+              onPressed: () => _permissionWithCustomPopup(),
             ),
             Padding(
               padding: const EdgeInsets.only(top: 50.0),
@@ -52,8 +53,8 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  _camera() async {
-    PermissionValidator permissionValidator = PermissionValidator(
+  _permissionRequest() async {
+    final permissionValidator = EasyPermissionValidator(
       context: context,
       appName: 'Easy Permission Validator',
     );
@@ -61,5 +62,93 @@ class _MyHomePageState extends State<MyHomePage> {
     if (result) {
       setState(() => _result = 'Permission accepted');
     }
+  }
+
+  _permissionWithCustomPopup() async {
+    EasyPermissionValidator permissionValidator = EasyPermissionValidator(
+      context: context,
+      appName: 'Easy Permission Validator',
+      customDialog: MyAmazingCustomPopup(),
+    );
+    var result = await permissionValidator.camera();
+    if (result) {
+      setState(() => _result = 'Permission accepted');
+    }
+  }
+}
+
+class MyAmazingCustomPopup extends StatefulWidget {
+  @override
+  _MyAmazingCustomPopupState createState() => _MyAmazingCustomPopupState();
+}
+
+class _MyAmazingCustomPopupState extends State<MyAmazingCustomPopup> {
+  @override
+  Widget build(BuildContext context) {
+    return WillPopScope(
+      onWillPop: () async => false,
+      child: Center(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 13.0),
+          child: SizedBox(
+            height: 300.0,
+            width: MediaQuery.of(context).size.width,
+            child: Card(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.all(Radius.circular(10.0)),
+              ),
+              color: Colors.white,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(vertical: 20.0),
+                child: Column(
+                  mainAxisSize: MainAxisSize.max,
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: <Widget>[
+                    Text(
+                      'Easy Permission Validator Demo',
+                      style: TextStyle(
+                        color: Colors.green,
+                        fontSize: 23.0,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    Icon(
+                      Icons.perm_camera_mic,
+                      size: 60.0,
+                      color: Colors.red,
+                    ),
+                    Row(
+                      mainAxisSize: MainAxisSize.max,
+                      mainAxisAlignment: MainAxisAlignment.spaceAround,
+                      children: <Widget>[
+                        FlatButton.icon(
+                          icon: Icon(Icons.cancel),
+                          onPressed: () => _closePopup(),
+                          label: Text('Cancel'),
+                        ),
+                        FlatButton.icon(
+                          icon: Icon(Icons.arrow_forward_ios),
+                          onPressed: () => _openPermissionSettings(),
+                          label: Text('Go To Settings'),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  _openPermissionSettings() async {
+    await PermissionHandler().openAppSettings();
+    _closePopup();
+  }
+
+  _closePopup() {
+    Navigator.of(context).pop();
   }
 }